Questionnaire Flattener (#12)
# Questionnaire Flattener

## :recycle: Current situation & Problem
The issue is
https://github.com/StanfordSpezi/SpeziDataPipelineTemplate/issues/11.
The flattener for questionnaire responses should be able to use
phoenix-generated surveys to convert the question and response IDs in
Firebase to text.


## :gear: Release Notes 

- Added the ability for a user to input Phoenix-generated JSON surveys
to ```flatten_fhir_resources``` and for the question and answer mappings
to be extracted. For the questions, I mapped the linkId to the text. For
answers, because they have both a system ID and choice code, I combined
the system id and code and used that as the mapping key.
https://github.com/StanfordSpezi/SpeziDataPipelineTemplate/blob/cca4c5d2501a1b66234c240f0066aa39dfc0fc6b/data_flattening/fhir_resources_flattener.py#L508
- Extracts the question and answer IDs from the individual responses and
then gets them from the survey mappings (if the answer is a
valueCoding). If it is not a valueCoding, it will just use the integer
or string value.
https://github.com/StanfordSpezi/SpeziDataPipelineTemplate/blob/cca4c5d2501a1b66234c240f0066aa39dfc0fc6b/data_flattening/fhir_resources_flattener.py#L541
- The flattener then generates a data frame with the _Date, UserID,
SurveyID, QuestionID, QuestionText,_ and _Answer_.
https://github.com/StanfordSpezi/SpeziDataPipelineTemplate/blob/cca4c5d2501a1b66234c240f0066aa39dfc0fc6b/data_flattening/fhir_resources_flattener.py#L573

## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).

---------

Co-authored-by: Vicky Bikia <vickybikia@gmail.com>
Co-authored-by: Paul Schmiedmayer <PSchmiedmayer@users.noreply.github.com>